+++ internal/pkg/pct/build.go
@@ -9,7 +9,12 @@ import (
 	"github.com/puppetlabs/pdkgo/internal/pkg/tar"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/afero"
+	"gopkg.in/yaml.v2"
 )
+
+type BuilderI interface {
+	Build(templatePath, targetDir string) (gzipArchiveFilePath string, err error)
+}
 
 type Builder struct {
 	Tar  tar.TarI
@@ -26,6 +31,11 @@ func (b *Builder) Build(templatePath, targetDir string) (gzipArchiveFilePath str
 	// Check if pct-config.yml exists
 	if _, err := b.AFS.Stat(filepath.Join(templatePath, "pct-config.yml")); os.IsNotExist(err) {
 		return "", fmt.Errorf("No 'pct-config.yml' found in %v", templatePath)
+	}
+
+	err = b.checkConfig(filepath.Join(templatePath, "pct-config.yml"))
+	if err != nil {
+		return "", fmt.Errorf("Invalid config: %v", err.Error())
 	}
 
 	// Check if content dir exists
@@ -56,4 +66,37 @@ func (b *Builder) Build(templatePath, targetDir string) (gzipArchiveFilePath str
 	}
 
 	return gzipArchiveFilePath, nil
+}
+
+func (b *Builder) checkConfig(configFile string) error {
+
+	fileBytes, err := b.AFS.ReadFile(configFile)
+	if err != nil {
+		return err
+	}
+
+	var info PuppetContentTemplateInfo
+	err = yaml.Unmarshal(fileBytes, &info)
+	if err != nil {
+		return err
+	}
+
+	msg := "The following attributes are missing in pct-config.yml:\n"
+	orig := msg
+	// These parts are essential for build and deployment.
+
+	if info.Template.Id == "" {
+		msg = msg + "  * id\n"
+	}
+	if info.Template.Author == "" {
+		msg = msg + "  * author\n"
+	}
+	if info.Template.Version == "" {
+		msg = msg + "  * version\n"
+	}
+	if msg != orig {
+		return fmt.Errorf(msg)
+	}
+
+	return nil
 }